+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
  * Change role to group in the Asset and Job workflows. (rudaporto)
  * Integrate workflow fix in briefy.common. (rudaporto)
  * BODY-62: Implement pagination. (ericof)
+ * LEICA-09: Improve Customers, Projects and Jobs import. Add service to run the import by API call. (rudaporto)
 
 1.1.0 (2016-10-04)
 ------------------